+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix inconsistent handling of invalid pcb.toml files between `pcb build` and `pcb publish`
+
 ## [0.3.20] - 2026-01-09
 
 ### Added